another static method override test
if both are static. method called will be the method of its own. so if you instantiated a Lion object. any method call with static override will call the Lions method. if Animal Object. it will call the Animal's static method.// If non static. always the child class. if instantiated as child
.